refactor Resolution and Alignment class
