+++ backend/src/main/java/com/bakdata/conquery/apiv1/forms/export_form/AbsoluteMode.java
@@ -15,7 +15,7 @@ import com.bakdata.conquery.models.auth.entities.User;
 import com.bakdata.conquery.models.common.Range;
 import com.bakdata.conquery.models.datasets.Dataset;
 import com.bakdata.conquery.models.forms.export.AbsExportGenerator;
-import com.bakdata.conquery.models.forms.util.DateContext;
+import com.bakdata.conquery.models.forms.util.Alignment;
 import com.bakdata.conquery.models.query.DateAggregationMode;
 import com.bakdata.conquery.models.query.QueryResolveContext;
 import com.bakdata.conquery.models.query.Visitable;
@@ -39,7 +39,7 @@ public class AbsoluteMode extends Mode {
 	}
 
 	@NotNull
-	private DateContext.Alignment alignmentHint = DateContext.Alignment.QUARTER;
+	private Alignment alignmentHint = Alignment.QUARTER;
 
 	@Override
 	public Query createSpecializedQuery(DatasetRegistry datasets, User user, Dataset submittedDataset) {
