+++ backend/src/main/java/com/bakdata/conquery/models/forms/managed/RelativeFormQueryPlan.java
@@ -8,6 +8,7 @@ import com.bakdata.conquery.apiv1.forms.export_form.ExportForm;
 import com.bakdata.conquery.models.common.CDateSet;
 import com.bakdata.conquery.models.error.ConqueryError;
 import com.bakdata.conquery.models.forms.util.DateContext;
+import com.bakdata.conquery.models.forms.util.Resolution;
 import com.bakdata.conquery.models.forms.util.ResultModifier;
 import com.bakdata.conquery.models.query.QueryExecutionContext;
 import com.bakdata.conquery.apiv1.query.concept.specific.temporal.TemporalSampler;
@@ -197,13 +198,13 @@ public class RelativeFormQueryPlan implements QueryPlan<MultilineEntityResult> {
 		if (featurePlan.getAggregatorSize() > 0 && outcomePlan.getAggregatorSize() > 0) {
 			// We have features and outcomes check if both have complete date ranges (they should be at the beginning of the list)
 			return contexts.size()>=2
-				&& contexts.get(0).getSubdivisionMode().equals(DateContext.Resolution.COMPLETE)
-				&& contexts.get(1).getSubdivisionMode().equals(DateContext.Resolution.COMPLETE)
+				&& contexts.get(0).getSubdivisionMode().equals(Resolution.COMPLETE)
+				&& contexts.get(1).getSubdivisionMode().equals(Resolution.COMPLETE)
 				&& !contexts.get(0).getFeatureGroup().equals(contexts.get(1).getFeatureGroup());
 		}
 		// Otherwise, if only features or outcomes are given check the first date context. The empty feature/outcome query
 		// will still return an empty result which will be merged with to a complete result.
-		return contexts.get(0).getSubdivisionMode().equals(DateContext.Resolution.COMPLETE);
+		return contexts.get(0).getSubdivisionMode().equals(Resolution.COMPLETE);
 	}
 
 	private FormQueryPlan createSubQuery(ArrayConceptQueryPlan subPlan, List<DateContext> contexts, FeatureGroup featureGroup) {
